merge: apply .rej files, ASCII comments pass, ffunicode UTF-16 guards
Apply rejected patch hunks across 4 files:
- driver_fat_shell.c: Unicode escape parsing (\uXXXX, \UXXXXXXXX)
- ffunicode.c: guard c16rtomb/mbrtoc16/towupper with __STDC_UTF_16__
- ff_exfat.c: remove internal __STDC_UTF_16__ guard (redundant with ff.h)
- aes_xts_impl.c: replace Unicode comment chars with ASCII

Batch cosmetic: replace Unicode ─/—/→ with ASCII --/--/-> in comments
across 31 files.

+++ backend/bklibkey.c
@@ -433,7 +433,7 @@ void add_key_to_keyslot(
    const bool    is_anonymous_key,
    const bool    is_allow_nolock,
    int *         ret_target_level) {
-#define SET_BIT(n, x, y) ((n & ~(1U << (x))) | ((y & 1) << (x)))
+#define SET_BIT(n, x, y) ((n & ~(1ULL << (x))) | (((uint64_t)(y) & 1) << (x)))
 
    uint8_t inited_key[HASHLEN];
